Ensure registration to moodle
We have a post-installation CLI task to queue, and then register the site.
This depends upon DNS being setup and available, hence an Adhoc task to
attempt to determine this (note, this can be broken if a site was
previously registered but now points to the wrong region).

An additional renderer prevents removal from the Moodle.net hub by removing
the button, and a hack in the registration index prevents the actual
removal if has been hacked in place.