MDL-68166 backup: Write backup files to localcache before saving
When backing up in Moodle we slowly build up our final backup file by
creating an initial file, and then adding each piece of data to it.

In a large backup, such as when backing up a large course, this can lead
to many writes to the same file which has a negative impact upon
performance, especially for clustered file systems such as NFS,
GlusterFS, and Ceph.

This change modifies the process so that the chunked writing of the
final backup file is to a file within the local cache directory. In a
Moodle cluster this is typically configured to be a directory local to
each node. This should have a much lower impact when writing a file in
chunks.

Once the entire file is created it is then moved to the final
destination. This final move should have a much lower performance hit
than writing the same file many times, especially for networked file
systems.

Where networked file systems are not in use, this change will have
little to no impact. Typically the localcachedir and the tempdir will be
on the same file system, and therefore an atomic rename will take place
at minimal cost.

Although there are more overall operations, they have a much lower cost
in configurations where performance is important.