NO-TASK: Allow H5P content types task to run on templates again
Now that we have a mechanism for propagating the H5P files to each region
there should never be a need to mess with the file system plugin to push
the files.

This means that there should never be a chance of template files pushing
things to S3.

If we run the content types task before making a site a template, the latest
content types will be installed and should be available immediatly for site.

The cron task will then just be keeping older sites up to date.